feat: support inexistant config and add lua file

+++ installAll.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+
 # Installing yay
-/InstallScripts/yay_download.sh
+InstallScripts/yay_download.sh
 if [[ $? != 0 ]]; then
     echo "Could not install yay, aborting..."
     exit 1
@@ -11,12 +13,26 @@ fi
 yay -S --noconfirm --needed trash-cli kitty firefox
 
 # Setting config files
-mv ~/.bashrc ~/.bashrc.bak #bashrc
+if [ -f ~/.bashrc ]; then #bashrc
+	mv ~/.bashrc ~/.bashrc.bak 
+fi
 ln -s ConfigFile/.bashrc ~/.bashrc
-mv ~/.config/starship.toml ~/.config/starship.toml.bak # catpuccine mocha starship 
+
+if [ -f ~/.config/starship.toml ]; then # catpuccine mocha starship 
+	mv ~/.config/starship.toml ~/.config/starship.toml.bak 
+fi
 ln -s ConfigFile/starship.toml ~/.config/starship.toml
-mv ~/.config/fastfetch/config.jsonc ~/.config/fastfetch/config.jsonc.bak # catpuccine mocha starship 
+
+if [ -f ~/.config/fastfetch/config.jsonc ]; then # custom fastfetch
+	mv ~/.config/fastfetch/config.jsonc ~/.config/fastfetch/config.jsonc.bak
+fi
 ln -s ConfigFile/fastfetchConfig.jsonc ~/.config/fastfetch/config.jsonc
+
+if [ -f ~/.config/nvim/init.lua]; then # custom lua
+	echo "dotfile($SCRIPT_DIR/ConfigFiles/init.lua)" >> ~/.config/nvim/init.lua
+else
+    ln -s ConfigFile/init.lua ~/.config/nvim/init.lua
+fi
 
 # Additional packages
 yay -S --noconfirm --needed vesktop tor rust devtoolbox onlyoffice isoimagewriter nextcloud-client beeper-v4-bin vlc kdeconnect gabutdm localsend
